+++ tests/integration/fixtures/relay.py
@@ -142,9 +142,7 @@ def relay(mini_sentry, random_port, background_process, config_dir, get_relay_bi
         dir = config_dir("relay")
         dir.join("config.yml").write(yaml.dump(default_opts))
 
-        output = subprocess.check_output(
-            relay_bin + ["-c", str(dir), "credentials", "generate"]
-        )
+        subprocess.check_output(relay_bin + ["-c", str(dir), "credentials", "generate"])
 
         # now that we have generated a credentials file get the details
         with open(dir.join("credentials.json")) as f: