+++ tests/integration/test_metrics.py
@@ -1,5 +1,4 @@
 import hashlib
-import time
 from collections import defaultdict
 from datetime import datetime, timedelta, timezone
 import json
@@ -64,7 +63,7 @@ def test_metrics(mini_sentry, relay):
     mini_sentry.add_basic_project_config(project_id)
 
     timestamp = int(datetime.now(tz=timezone.utc).timestamp())
-    metrics_payload = f"transactions/foo:42|c\ntransactions/bar:17|c"
+    metrics_payload = "transactions/foo:42|c\ntransactions/bar:17|c"
     relay.send_metrics(project_id, metrics_payload, timestamp)
 
     envelope = mini_sentry.captured_events.get(timeout=3)
@@ -100,7 +99,7 @@ def test_metrics_backdated(mini_sentry, relay):
     mini_sentry.add_basic_project_config(project_id)
 
     timestamp = int(datetime.now(tz=timezone.utc).timestamp()) - 24 * 60 * 60
-    metrics_payload = f"transactions/foo:42|c"
+    metrics_payload = "transactions/foo:42|c"
     relay.send_metrics(project_id, metrics_payload, timestamp)
 
     envelope = mini_sentry.captured_events.get(timeout=2)
@@ -152,7 +151,7 @@ def test_metrics_partition_key(mini_sentry, relay, flush_partitions, expected_he
     )
 
     timestamp = 999994711
-    metrics_payload = f"transactions/foo:42|c"
+    metrics_payload = "transactions/foo:42|c"
     relay.send_metrics(project_id, metrics_payload, timestamp)
 
     mini_sentry.captured_events.get(timeout=3)
@@ -172,7 +171,7 @@ def test_metrics_with_processing(mini_sentry, relay_with_processing, metrics_con
     mini_sentry.add_full_project_config(project_id)
 
     timestamp = int(datetime.now(tz=timezone.utc).timestamp())
-    metrics_payload = f"transactions/foo:42|c\ntransactions/bar@second:17|c"
+    metrics_payload = "transactions/foo:42|c\ntransactions/bar@second:17|c"
     relay.send_metrics(project_id, metrics_payload, timestamp)
 
     metrics = metrics_by_name(metrics_consumer, 2)
@@ -243,7 +242,7 @@ def test_metrics_with_sharded_kafka(
     m1 = metrics_consumer(topic="metrics-1")
     m2 = metrics_consumer(topic="metrics-2")
 
-    metrics_payload = f"transactions/foo:42|c\ntransactions/bar@second:17|c"
+    metrics_payload = "transactions/foo:42|c\ntransactions/bar@second:17|c"
 
     timestamp = int(datetime.now(tz=timezone.utc).timestamp())
     relay.send_metrics(project_id, metrics_payload, timestamp)
@@ -295,10 +294,10 @@ def test_metrics_full(mini_sentry, relay, relay_with_processing, metrics_consume
 
     # Send two events to downstream and one to upstream
     timestamp = int(datetime.now(tz=timezone.utc).timestamp())
-    downstream.send_metrics(project_id, f"transactions/foo:7|c", timestamp)
-    downstream.send_metrics(project_id, f"transactions/foo:5|c", timestamp)
+    downstream.send_metrics(project_id, "transactions/foo:7|c", timestamp)
+    downstream.send_metrics(project_id, "transactions/foo:5|c", timestamp)
 
-    upstream.send_metrics(project_id, f"transactions/foo:3|c", timestamp)
+    upstream.send_metrics(project_id, "transactions/foo:3|c", timestamp)
 
     metric, _ = metrics_consumer.get_metric(timeout=6)
     metric.pop("timestamp")
@@ -579,6 +578,7 @@ def test_transaction_metrics(
     project_id = 42
     mini_sentry.add_full_project_config(project_id)
     config = mini_sentry.project_configs[project_id]["config"]
+
     timestamp = datetime.now(tz=timezone.utc)
 
     if extract_metrics:
@@ -1035,17 +1035,17 @@ def test_graceful_shutdown(mini_sentry, relay):
 
     # Backdated metric will be flushed immediately due to debounce delay
     past_timestamp = timestamp - 1000
-    metrics_payload = f"transactions/foo:42|c"
+    metrics_payload = "transactions/foo:42|c"
     relay.send_metrics(project_id, metrics_payload, past_timestamp)
 
     # Future timestamp will not be flushed regularly, only through force flush
-    metrics_payload = f"transactions/bar:17|c"
+    metrics_payload = "transactions/bar:17|c"
     future_timestamp = timestamp + 30
     relay.send_metrics(project_id, metrics_payload, future_timestamp)
     relay.shutdown(sig=signal.SIGTERM)
 
     # Try to send another metric (will be rejected)
-    metrics_payload = f"transactions/zap:666|c"
+    metrics_payload = "transactions/zap:666|c"
     with pytest.raises(requests.ConnectionError):
         relay.send_metrics(project_id, metrics_payload, timestamp)
 
@@ -1200,7 +1200,7 @@ def test_span_metrics(
         for metric, headers in metrics
         if metric["name"].startswith("spans", 2)
     ]
-    assert len(span_metrics) == 3
+    assert len(span_metrics) == 2
     for metric, headers in span_metrics:
         assert headers == [("namespace", b"spans")]
         assert metric["tags"]["span.description"] == expected_description
@@ -1283,7 +1283,7 @@ def test_span_metrics_secondary_aggregator(
         },
         "spans": [
             {
-                "description": "this_is_too_long",
+                "description": "SELECT %s FROM foo",
                 "op": "db",
                 "parent_span_id": "8f5a2b8768cafb4e",
                 "span_id": "bd429c44b67a3eb4",
@@ -1315,6 +1315,7 @@ def test_span_metrics_secondary_aggregator(
                         "bucket_interval": 1,
                         "initial_delay": 0,
                         "debounce_delay": 0,
+                        "max_tag_value_length": 10,
                     },
                 }
             ],
@@ -1330,20 +1331,21 @@ def test_span_metrics_secondary_aggregator(
     span_metrics = [
         (metric, headers)
         for metric, headers in metrics
-        if metric["name"] == "d:spans/duration@millisecond"
+        if metric["name"] == "d:spans/exclusive_time@millisecond"
     ]
     assert span_metrics == [
         (
             {
-                "name": "d:spans/duration@millisecond",
+                "name": "d:spans/exclusive_time@millisecond",
                 "org_id": 1,
                 "project_id": 42,
                 "retention_days": 90,
                 "tags": {
+                    "span.action": "SELECT",
+                    "span.description": "SELECT %s*",
                     "span.category": "db",
                     "span.module": "db",
                     "span.op": "db",
-                    "transaction": "/organizations/:orgId/performance/:eventSlug/",
                 },
                 "timestamp": int(timestamp.timestamp()),
                 "type": "d",